+++ damaba/src/test/kotlin/com/damaba/damaba/application/port/inbound/promotion/PostPromotionUseCaseCommandTest.kt
@@ -8,8 +8,8 @@ import com.damaba.damaba.domain.promotion.constant.PromotionType
 import com.damaba.damaba.domain.region.Region
 import com.damaba.damaba.util.RandomTestUtils.Companion.randomLong
 import com.damaba.damaba.util.RandomTestUtils.Companion.randomString
-import com.damaba.damaba.util.TestFixture.createAddress
-import com.damaba.damaba.util.TestFixture.createFile
+import com.damaba.damaba.util.fixture.AddressFixture.createAddress
+import com.damaba.damaba.util.fixture.FileFixture.createImage
 import org.assertj.core.api.Assertions.assertThat
 import org.assertj.core.api.Assertions.catchThrowable
 import org.junit.jupiter.params.ParameterizedTest
@@ -71,7 +71,7 @@ class PostPromotionUseCaseCommandTest {
         @JvmStatic
         fun invalidImageSizeProvider() = listOf(
             Arguments.of(emptyList<File>()),
-            Arguments.of(List(11) { createFile() }),
+            Arguments.of(List(11) { createImage() }),
         )
 
         private fun createCommand(
@@ -80,7 +80,7 @@ class PostPromotionUseCaseCommandTest {
             content: String = "Valid content",
             promotionType: PromotionType = PromotionType.EVENT,
             eventType: EventType = EventType.FREE,
-            images: List<File> = List(3) { createFile() },
+            images: List<File> = List(3) { createImage() },
             activeRegions: Set<Region> = setOf(Region("서울", "강남구")),
         ) = PostPromotionUseCase.Command(
             authorId = randomLong(),
