+++ damaba/src/test/kotlin/com/damaba/damaba/util/fixture/PhotographerFixture.kt
@@ -1,0 +1,82 @@
+package com.damaba.damaba.util.fixture
+
+import com.damaba.common_file.domain.Image
+import com.damaba.damaba.adapter.outbound.photographer.BusinessScheduleJpaEmbeddable
+import com.damaba.damaba.adapter.outbound.photographer.PhotographerActiveRegionJpaEntity
+import com.damaba.damaba.adapter.outbound.photographer.PhotographerAddressJpaEmbeddable
+import com.damaba.damaba.adapter.outbound.photographer.PhotographerJpaEntity
+import com.damaba.damaba.adapter.outbound.photographer.PhotographerPortfolioImageJpaEntity
+import com.damaba.damaba.domain.common.Address
+import com.damaba.damaba.domain.common.PhotographyType
+import com.damaba.damaba.domain.photographer.BusinessSchedule
+import com.damaba.damaba.domain.photographer.Photographer
+import com.damaba.damaba.domain.region.Region
+import com.damaba.damaba.util.RandomTestUtils.Companion.randomLong
+import com.damaba.damaba.util.RandomTestUtils.Companion.randomString
+import com.damaba.damaba.util.RandomTestUtils.Companion.randomUrl
+import com.damaba.damaba.util.fixture.AddressFixture.createAddress
+import com.damaba.user.domain.user.constant.Gender
+import com.damaba.user.domain.user.constant.LoginType
+import com.damaba.user.domain.user.constant.UserRoleType
+import com.damaba.user.domain.user.constant.UserType
+
+object PhotographerFixture {
+    fun createPhotographer(
+        id: Long = randomLong(),
+        type: UserType = UserType.PHOTOGRAPHER,
+        roles: Set<UserRoleType> = setOf(UserRoleType.USER),
+        oAuthLoginUid: String = randomString(),
+        loginType: LoginType = LoginType.KAKAO,
+        nickname: String = randomString(len = 7),
+        profileImage: Image = Image(randomString(), randomUrl()),
+        gender: Gender = Gender.MALE,
+        instagramId: String = randomString(len = 30),
+        mainPhotographyTypes: Set<PhotographyType> = setOf(PhotographyType.PROFILE),
+        contactLink: String? = randomString(),
+        description: String? = randomString(),
+        address: Address? = createAddress(),
+        businessSchedule: BusinessSchedule? = null,
+        portfolio: List<Image> = emptyList(),
+        activeRegions: Set<Region> = emptySet(),
+    ) = Photographer(
+        id = id,
+        type = type,
+        roles = roles,
+        loginType = loginType,
+        oAuthLoginUid = oAuthLoginUid,
+        nickname = nickname,
+        profileImage = profileImage,
+        gender = gender,
+        instagramId = instagramId,
+        mainPhotographyTypes = mainPhotographyTypes,
+        contactLink = contactLink,
+        description = description,
+        address = address,
+        businessSchedule = businessSchedule,
+        portfolio = portfolio,
+        activeRegions = activeRegions,
+    )
+
+    fun createPhotographerJpaEntity(
+        id: Long = randomLong(),
+        mainPhotographyTypes: Set<PhotographyType> = setOf(PhotographyType.PROFILE),
+        contactLink: String? = randomString(),
+        description: String? = randomString(),
+        address: PhotographerAddressJpaEmbeddable? = null,
+        businessSchedule: BusinessScheduleJpaEmbeddable? = null,
+        portfolio: List<PhotographerPortfolioImageJpaEntity> = emptyList(),
+        activeRegions: Set<PhotographerActiveRegionJpaEntity> = emptySet(),
+    ): PhotographerJpaEntity {
+        val photographerJpaEntity = PhotographerJpaEntity(
+            userId = id,
+            mainPhotographyTypes = mainPhotographyTypes,
+            contactLink = contactLink,
+            description = description,
+            address = address,
+            businessSchedule = businessSchedule,
+        )
+        photographerJpaEntity.addPortfolioImages(portfolio)
+        photographerJpaEntity.activeRegions.addAll(activeRegions)
+        return photographerJpaEntity
+    }
+}
