+++ damaba/src/test/kotlin/com/damaba/damaba/util/fixture/RegionFixture.kt
@@ -1,0 +1,24 @@
+package com.damaba.damaba.util.fixture
+
+import com.damaba.damaba.adapter.inbound.region.dto.RegionRequest
+import com.damaba.damaba.domain.region.Region
+import com.damaba.damaba.domain.region.RegionGroup
+import com.damaba.damaba.util.RandomTestUtils.Companion.generateRandomList
+import com.damaba.damaba.util.RandomTestUtils.Companion.randomString
+
+object RegionFixture {
+    fun createRegionRequest(
+        category: String = randomString(),
+        name: String = randomString(),
+    ) = RegionRequest(category, name)
+
+    fun createRegion(category: String = randomString(), name: String = randomString()) = Region(category, name)
+
+    private fun createRegionGroup(): RegionGroup = RegionGroup(
+        category = randomString(),
+        regions = generateRandomList(maxSize = 10) { randomString() },
+    )
+
+    fun createRegionGroups(): List<RegionGroup> =
+        generateRandomList(maxSize = 10) { createRegionGroup() }
+}
