+++ damaba/src/test/kotlin/com/damaba/damaba/util/fixture/UserFixture.kt
@@ -1,0 +1,57 @@
+package com.damaba.damaba.util.fixture
+
+import com.damaba.common_file.domain.Image
+import com.damaba.damaba.util.RandomTestUtils.Companion.randomLong
+import com.damaba.damaba.util.RandomTestUtils.Companion.randomString
+import com.damaba.damaba.util.RandomTestUtils.Companion.randomUrl
+import com.damaba.user.adapter.outbound.user.UserJpaEntity
+import com.damaba.user.adapter.outbound.user.UserProfileImageJpaEmbeddable
+import com.damaba.user.domain.user.User
+import com.damaba.user.domain.user.constant.Gender
+import com.damaba.user.domain.user.constant.LoginType
+import com.damaba.user.domain.user.constant.UserRoleType
+import com.damaba.user.domain.user.constant.UserType
+
+object UserFixture {
+    fun createUser(
+        id: Long = randomLong(),
+        type: UserType = UserType.USER,
+        roles: Set<UserRoleType> = setOf(UserRoleType.USER),
+        oAuthLoginUid: String = randomString(),
+        loginType: LoginType = LoginType.KAKAO,
+        nickname: String = randomString(len = 7),
+        profileImage: Image = Image(randomString(), randomUrl()),
+        gender: Gender = Gender.MALE,
+        instagramId: String = randomString(len = 30),
+    ): User = User(
+        id = id,
+        type = type,
+        roles = roles,
+        oAuthLoginUid = oAuthLoginUid,
+        loginType = loginType,
+        nickname = nickname,
+        profileImage = profileImage,
+        gender = gender,
+        instagramId = instagramId,
+    )
+
+    fun createUserJpaEntity(
+        type: UserType = UserType.USER,
+        roles: Set<UserRoleType> = setOf(UserRoleType.USER),
+        oAuthLoginUid: String = randomString(),
+        loginType: LoginType = LoginType.KAKAO,
+        nickname: String = randomString(len = 7),
+        profileImage: UserProfileImageJpaEmbeddable = UserProfileImageJpaEmbeddable(randomString(), randomUrl()),
+        gender: Gender = Gender.MALE,
+        instagramId: String = randomString(len = 30),
+    ) = UserJpaEntity(
+        type = type,
+        roles = roles,
+        oAuthLoginUid = oAuthLoginUid,
+        loginType = loginType,
+        nickname = nickname,
+        profileImage = profileImage,
+        gender = gender,
+        instagramId = instagramId,
+    )
+}
